src/architecture-review/requirements.txt
pytest>=7.0.0
pyyaml>=6.0
jsonschema>=4.0.0
